changes revert for periodic sync
