+++ android/deps.gradle
@@ -42,7 +42,7 @@ versions.robolectric = '4.4'
 versions.room = '2.3.0'
 versions.spotless = '5.11.0'
 versions.truth = '1.0.1'
-versions.work = '2.6.0'
+versions.work = '2.3.4'
 versions.json_tools = '1.13'
 versions.kotlin_coveralls = '2.12.0'
 versions.jacoco_tool = '0.8.7'
