+++ android/engine/build.gradle
@@ -10,23 +10,6 @@ plugins {
 
 apply from: '../jacoco.gradle'
 
-
-Properties props = new Properties()
-
-def arr = ["FHIR_BASE_URL", "OAUTH_BASE_URL", "OAUTH_CIENT_ID", "OAUTH_CLIENT_SECRET", "OAUTH_SCOPE"]
-if (rootProject.file("local.properties").exists()) {
-    props.load(rootProject.file("local.properties").newDataInputStream())
-
-    arr.each { prop ->
-        project.ext.set(prop, props.getProperty(prop, (prop == "FHIR_BASE_URL") ? "https://sample.url/fhir/" : "sample"))
-    }
-} else {
-    println("local.properties does not exist. The following values are required " + arr.join(", "))
-    arr.each { prop ->
-        project.ext.set(prop, props.getProperty(prop, (prop.contains("URL")) ? "https://sample.url/fhir/" : "sample_" + prop))
-    }
-}
-
 android {
 
     compileSdkVersion 30
@@ -38,13 +21,6 @@ android {
         targetSdkVersion 30
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles "consumer-rules.pro"
-
-        buildConfigField("boolean", 'SKIP_AUTH_CHECK', "false")
-        buildConfigField("String", 'FHIR_BASE_URL', "\"${FHIR_BASE_URL}\"")
-        buildConfigField("String", 'OAUTH_BASE_URL', "\"${OAUTH_BASE_URL}\"")
-        buildConfigField("String", 'OAUTH_CIENT_ID', "\"${OAUTH_CIENT_ID}\"")
-        buildConfigField("String", 'OAUTH_CLIENT_SECRET', "\"${OAUTH_CLIENT_SECRET}\"")
-        buildConfigField("String", 'OAUTH_SCOPE', "\"${OAUTH_SCOPE}\"")
     }
 
     buildTypes {
@@ -153,8 +129,6 @@ dependencies {
     //Hilt - Dependency Injection
     implementation "com.google.dagger:hilt-android:$hiltVersion"
     kapt "com.google.dagger:hilt-compiler:$hiltVersion"
-    implementation("androidx.hilt:hilt-work:1.0.0")
-    kapt("androidx.hilt:hilt-compiler:1.0.0")
 
     api 'androidx.lifecycle:lifecycle-livedata-ktx:2.3.1'
     api 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.1'
@@ -171,6 +145,7 @@ dependencies {
     api("androidx.paging:paging-compose:1.0.0-alpha12")
     api("androidx.activity:activity-compose:1.3.1")
     api 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.2.2'
+    api("androidx.work:work-runtime-ktx:2.3.4")
 
     def coroutineVersion = '1.5.1'
     api("org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutineVersion")
