+++ android/engine/src/main/java/org/smartregister/fhircore/engine/auth/AccountAuthenticator.kt
@@ -34,12 +34,12 @@ import java.util.Locale
 import javax.inject.Inject
 import javax.inject.Singleton
 import okhttp3.ResponseBody
-import org.smartregister.fhircore.engine.R
 import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
 import org.smartregister.fhircore.engine.data.remote.auth.OAuthService
 import org.smartregister.fhircore.engine.data.remote.model.response.OAuthResponse
 import org.smartregister.fhircore.engine.ui.appsetting.AppSettingActivity
 import org.smartregister.fhircore.engine.ui.login.LoginActivity
+import org.smartregister.fhircore.engine.util.APP_ID_CONFIG
 import org.smartregister.fhircore.engine.util.SecureSharedPreference
 import org.smartregister.fhircore.engine.util.SharedPreferencesHelper
 import org.smartregister.fhircore.engine.util.toSha1
@@ -299,6 +299,7 @@ constructor(
       secureSharedPreference.deleteCredentials()
       launchScreen(AppSettingActivity::class.java)
     }
+    sharedPreference.remove(APP_ID_CONFIG)
   }
 
   fun launchLoginScreen() {
@@ -320,7 +321,7 @@ constructor(
 
   fun getLoginActivityClass(): Class<*> = LoginActivity::class.java
 
-  fun getAccountType(): String = context.getString(R.string.authenticator_account_type)
+  fun getAccountType(): String = configurationRegistry.authConfiguration.accountType
 
   fun clientSecret(): String = configurationRegistry.authConfiguration.clientSecret
