+++ android/engine/src/main/java/org/smartregister/fhircore/engine/configuration/app/AuthConfiguration.kt
@@ -22,5 +22,6 @@ data class AuthConfiguration(
   var fhirServerBaseUrl: String,
   var clientId: String,
   var clientSecret: String,
+  var accountType: String,
   var scope: String = "openid"
 )
