--- android/engine/src/main/java/org/smartregister/fhircore/engine/di/CommonModule.kt
@@ -1,45 +1,0 @@
-/*
- * Copyright 2021 Ona Systems, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.smartregister.fhircore.engine.di
-
-import android.accounts.AccountManager
-import android.content.Context
-import com.google.android.fhir.FhirEngine
-import com.google.android.fhir.FhirEngineProvider
-import com.google.android.fhir.sync.Sync
-import com.google.android.fhir.sync.SyncJob
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-import org.hl7.fhir.r4.context.SimpleWorkerContext
-import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
-import org.smartregister.fhircore.engine.data.remote.fhir.resource.FhirResourceDataSource
-import org.smartregister.fhircore.engine.sync.SyncBroadcaster
-import org.smartregister.fhircore.engine.util.SecureSharedPreference
-
-@InstallIn(SingletonComponent::class)
-@Module
-class CommonModule {
-
-  @Singleton
-  @Provides
-  fun provideSecureSharedPreference(@ApplicationContext context: Context):SecureSharedPreference =
-    SecureSharedPreference(context)
-}
