+++ android/engine/src/main/java/org/smartregister/fhircore/engine/di/EngineModule.kt
@@ -62,4 +62,9 @@ class EngineModule {
     )
 
   @Singleton @Provides fun provideWorkerContextProvider() = SimpleWorkerContext()
+
+  @Singleton
+  @Provides
+  fun provideApplicationManager(@ApplicationContext context: Context): AccountManager =
+    AccountManager.get(context)
 }
