+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/login/LoginActivity.kt
@@ -26,7 +26,6 @@ import javax.inject.Inject
 import org.smartregister.fhircore.engine.configuration.AppConfigClassification
 import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
 import org.smartregister.fhircore.engine.configuration.app.ApplicationConfiguration
-import org.smartregister.fhircore.engine.configuration.app.ConfigService
 import org.smartregister.fhircore.engine.configuration.view.ConfigurableComposableView
 import org.smartregister.fhircore.engine.configuration.view.LoginViewConfiguration
 import org.smartregister.fhircore.engine.sync.SyncBroadcaster
@@ -42,8 +41,6 @@ class LoginActivity :
 
   @Inject lateinit var configurationRegistry: ConfigurationRegistry
 
-  @Inject lateinit var configService: ConfigService
-
   @Inject lateinit var syncBroadcaster: SyncBroadcaster
 
   private val loginViewModel by viewModels<LoginViewModel>()
@@ -55,11 +52,10 @@ class LoginActivity :
       navigateToHome.observe(
         this@LoginActivity,
         {
+          syncBroadcaster.runSync()
           if (loginViewModel.loginViewConfiguration.value?.enablePin == true) {
             loginService.navigateToPinLogin(goForSetup = true)
           } else {
-            //TODO???????????????????
-            configService.schedulePeriodicSync(syncBroadcaster.syncJob, configurationRegistry, syncBroadcaster)
             loginService.navigateToHome()
           }
         }
