+++ android/engine/src/main/res/values/strings.xml
@@ -1,6 +1,4 @@
 <resources>
-    <string name="authenticator_account_type" translatable="false">org.smartregister.fhircore</string>
-
     <string name="sync">Sync</string>
     <string name="last_sync_timestamp">Last: %1$s</string>
     <string name="language">Language</string>
