+++ android/quest/build.gradle
@@ -12,6 +12,22 @@ plugins {
 
 apply from: '../jacoco.gradle'
 
+Properties props = new Properties()
+
+def arr = ["FHIR_BASE_URL", "OAUTH_BASE_URL", "OAUTH_CIENT_ID", "OAUTH_CLIENT_SECRET", "OAUTH_SCOPE"]
+if (rootProject.file("local.properties").exists()) {
+    props.load(rootProject.file("local.properties").newDataInputStream())
+
+    arr.each { prop ->
+        project.ext.set(prop, props.getProperty(prop, (prop == "FHIR_BASE_URL") ? "https://sample.url/fhir/" : "sample"))
+    }
+} else {
+    println("local.properties does not exist. The following values are required " + arr.join(", "))
+    arr.each { prop ->
+        project.ext.set(prop, props.getProperty(prop, (prop.contains("URL")) ? "https://sample.url/fhir/" : "sample_" + prop))
+    }
+}
+
 android {
     compileSdkVersion 30
     compileSdk 30
@@ -27,6 +43,13 @@ android {
         versionCode 3
         versionName "0.0.3"
         multiDexEnabled true
+
+        buildConfigField("boolean", 'SKIP_AUTH_CHECK', "false")
+        buildConfigField("String", 'FHIR_BASE_URL', "\"${FHIR_BASE_URL}\"")
+        buildConfigField("String", 'OAUTH_BASE_URL', "\"${OAUTH_BASE_URL}\"")
+        buildConfigField("String", 'OAUTH_CIENT_ID', "\"${OAUTH_CIENT_ID}\"")
+        buildConfigField("String", 'OAUTH_CLIENT_SECRET', "\"${OAUTH_CLIENT_SECRET}\"")
+        buildConfigField("String", 'OAUTH_SCOPE', "\"${OAUTH_SCOPE}\"")
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         minSdkVersion 21
@@ -108,6 +131,7 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.6.0'
     implementation 'androidx.appcompat:appcompat:1.3.1'
     implementation 'com.google.android.material:material:1.4.0'
+    implementation 'androidx.work:work-runtime-ktx:2.6.0'
     implementation 'androidx.ui:ui-foundation:0.1.0-dev03'
     implementation deps.lifecycle.viewmodel
     implementation deps.work.runtime
