+++ android/quest/src/main/java/org/smartregister/fhircore/quest/QuestConfigService.kt
@@ -93,4 +93,13 @@ constructor(
     }
     return searchParameters
   }
+
+  override fun provideAuthConfiguration() =
+    AuthConfiguration(
+      fhirServerBaseUrl = BuildConfig.FHIR_BASE_URL,
+      oauthServerBaseUrl = BuildConfig.OAUTH_BASE_URL,
+      clientId = BuildConfig.OAUTH_CIENT_ID,
+      clientSecret = BuildConfig.OAUTH_CLIENT_SECRET,
+      accountType = context.getString(R.string.authenticator_account_type)
+    )
 }
